Reinsert Original Data
