+++ src/components/AggregatedTranscriptions/components/TranscriptionTable/Flags.js
@@ -1,0 +1,23 @@
+import React from 'react'
+import { Box } from 'grommet'
+import styled from 'styled-components'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { faCircle, faStar } from '@fortawesome/free-solid-svg-icons'
+
+const StyledFontAwesomeIcon = styled(FontAwesomeIcon)`
+  height: 0.5em;
+`
+
+export default function Flags({ datum }) {
+  return (
+    <Box direction='row'>
+      <Box align='center' direction='row' justify='center' width='3em'>
+        {datum.reviewed && <StyledFontAwesomeIcon color='green' icon={faCircle} />}
+        {datum.flagged && <StyledFontAwesomeIcon color='tomato' icon={faCircle} />}
+      </Box>
+      <Box width='1em'>
+        {datum.goldStandard && <FontAwesomeIcon color='gold' icon={faStar} size='xs' />}
+      </Box>
+    </Box>
+  )
+}
